add password support
